refactored to map.js, took  out toolbar into its own component
